--- a/doc/modules/changes/20240606_dannberg
+++ b/doc/modules/changes/20240606_dannberg
@@ -1,0 +1,5 @@
+Changed: The grain size material model now uses SUNDIALs
+to compute the grain size change (rather than a hand-
+written forward Euler scheme).
+<br>
+(Juliane Dannberg, 2024/06/06)