--- a/tests/grain_size_growth_log/screen-output
+++ b/tests/grain_size_growth_log/screen-output
@@ -15,45 +15,45 @@ Number of degrees of freedom: 17,989 (8,450+1,089+4,225+4,225)
 
 *** Timestep 1:  t=15625 years, dt=15625 years
    Solving temperature system... 0 iterations.
-   Solving grain_size system ... 14 iterations.
-   Rebuilding Stokes preconditioner...
-   Solving Stokes system... 0+0 iterations.
-
-   Postprocessing:
-     Compositions min/max/mass: 9.431/9.433/9.431e+10
-     Temperature min/avg/max:   1600 K, 1600 K, 1600 K
-     RMS, max velocity:         0.1 m/year, 0.1 m/year
-
-*** Timestep 2:  t=31250 years, dt=15625 years
-   Solving temperature system... 0 iterations.
    Solving grain_size system ... 13 iterations.
    Rebuilding Stokes preconditioner...
    Solving Stokes system... 0+0 iterations.
 
    Postprocessing:
-     Compositions min/max/mass: 9.428/9.433/9.428e+10
+     Compositions min/max/mass: 9.433/9.434/9.434e+10
+     Temperature min/avg/max:   1600 K, 1600 K, 1600 K
+     RMS, max velocity:         0.1 m/year, 0.1 m/year
+
+*** Timestep 2:  t=31250 years, dt=15625 years
+   Solving temperature system... 0 iterations.
+   Solving grain_size system ... 10 iterations.
+   Rebuilding Stokes preconditioner...
+   Solving Stokes system... 0+0 iterations.
+
+   Postprocessing:
+     Compositions min/max/mass: 9.433/9.435/9.434e+10
      Temperature min/avg/max:   1600 K, 1600 K, 1600 K
      RMS, max velocity:         0.1 m/year, 0.1 m/year
 
 *** Timestep 3:  t=46875 years, dt=15625 years
    Solving temperature system... 0 iterations.
-   Solving grain_size system ... 14 iterations.
+   Solving grain_size system ... 11 iterations.
    Rebuilding Stokes preconditioner...
    Solving Stokes system... 0+0 iterations.
 
    Postprocessing:
-     Compositions min/max/mass: 9.425/9.433/9.425e+10
+     Compositions min/max/mass: 9.433/9.436/9.434e+10
      Temperature min/avg/max:   1600 K, 1600 K, 1600 K
      RMS, max velocity:         0.1 m/year, 0.1 m/year
 
 *** Timestep 4:  t=50000 years, dt=3125.02 years
    Solving temperature system... 0 iterations.
-   Solving grain_size system ... 5 iterations.
+   Solving grain_size system ... 6 iterations.
    Rebuilding Stokes preconditioner...
    Solving Stokes system... 0+0 iterations.
 
    Postprocessing:
-     Compositions min/max/mass: 9.424/9.433/9.425e+10
+     Compositions min/max/mass: 9.433/9.436/9.434e+10
      Temperature min/avg/max:   1600 K, 1600 K, 1600 K
      RMS, max velocity:         0.1 m/year, 0.1 m/year