--- a/tests/grain_size_growth_pinned/screen-output
+++ b/tests/grain_size_growth_pinned/screen-output
@@ -20,7 +20,7 @@ Number of degrees of freedom: 17,989 (8,450+1,089+4,225+4,225)
    Solving Stokes system... 0+0 iterations.
 
    Postprocessing:
-     Compositions min/max/mass: 8e-05/8.913e-05/8.897e+05
+     Compositions min/max/mass: 8e-05/8.907e-05/8.89e+05
      Temperature min/avg/max:   1600 K, 1600 K, 1600 K
      RMS, max velocity:         0.1 m/year, 0.1 m/year
 
@@ -31,18 +31,18 @@ Number of degrees of freedom: 17,989 (8,450+1,089+4,225+4,225)
    Solving Stokes system... 0+0 iterations.
 
    Postprocessing:
-     Compositions min/max/mass: 8e-05/9.364e-05/9.332e+05
+     Compositions min/max/mass: 8e-05/9.356e-05/9.324e+05
      Temperature min/avg/max:   1600 K, 1600 K, 1600 K
      RMS, max velocity:         0.1 m/year, 0.1 m/year
 
 *** Timestep 3:  t=46875 years, dt=15625 years
    Solving temperature system... 0 iterations.
-   Solving grain_size system ... 17 iterations.
+   Solving grain_size system ... 18 iterations.
    Rebuilding Stokes preconditioner...
    Solving Stokes system... 0+0 iterations.
 
    Postprocessing:
-     Compositions min/max/mass: 8e-05/9.662e-05/9.615e+05
+     Compositions min/max/mass: 8e-05/9.653e-05/9.607e+05
      Temperature min/avg/max:   1600 K, 1600 K, 1600 K
      RMS, max velocity:         0.1 m/year, 0.1 m/year
 
@@ -53,7 +53,7 @@ Number of degrees of freedom: 17,989 (8,450+1,089+4,225+4,225)
    Solving Stokes system... 0+0 iterations.
 
    Postprocessing:
-     Compositions min/max/mass: 8e-05/9.715e-05/9.666e+05
+     Compositions min/max/mass: 8e-05/9.707e-05/9.658e+05
      Temperature min/avg/max:   1600 K, 1600 K, 1600 K
      RMS, max velocity:         0.1 m/year, 0.1 m/year