--- a/tests/grain_size_ridge/screen-output
+++ b/tests/grain_size_ridge/screen-output
@@ -70,20 +70,20 @@ Number of degrees of freedom: 57,556 (21,186+2,737+10,593+23,040)
    Solving temperature system... 7 iterations.
    Advecting particles...  done.
    Solving Stokes system... 1+0 iterations.
-      Relative nonlinear residuals (temperature, compositional fields, Stokes system): 2.07544e-05, 0, 0.000124185
-      Relative nonlinear residual (total system) after nonlinear iteration 1: 0.000124185
+      Relative nonlinear residuals (temperature, compositional fields, Stokes system): 2.07544e-05, 0, 0.000124194
+      Relative nonlinear residual (total system) after nonlinear iteration 1: 0.000124194
 
    Solving temperature system... 5 iterations.
    Advecting particles...  done.
    Solving Stokes system... 1+0 iterations.
-      Relative nonlinear residuals (temperature, compositional fields, Stokes system): 1.49196e-07, 0, 7.78945e-05
-      Relative nonlinear residual (total system) after nonlinear iteration 2: 7.78945e-05
+      Relative nonlinear residuals (temperature, compositional fields, Stokes system): 1.45943e-07, 0, 7.78278e-05
+      Relative nonlinear residual (total system) after nonlinear iteration 2: 7.78278e-05
 
 
    Postprocessing:
-     Compositions min/max/mass:          0.0005804/0.005/7.93e+09
+     Compositions min/max/mass:          0.0006276/0.005/7.962e+09
      RMS, max velocity:                  0.0355 m/year, 0.118 m/year
-     Mass fluxes through boundary parts: 0 kg/yr, 4.133e+07 kg/yr, -1.695e+07 kg/yr, 0 kg/yr
+     Mass fluxes through boundary parts: 0 kg/yr, 4.133e+07 kg/yr, -1.688e+07 kg/yr, 0 kg/yr
      Number of advected particles:       499998
 
 Termination requested by criterion: end time