This PR changes the grain size material model to use SUNDIALs to compute the grain size change. 
This is still about 30% slower than the original implementation. I am also not sure how to set all the ARKode parameters. The SUNDIALs documentation says that it tries 7 times to decrease the timestep if the error is too large, and that was not enough in some of the cases I tried, so I'd like to increase this number (rather than decreasing the initial step size). 

* [x] I have followed the [instructions for indenting my code](../blob/main/CONTRIBUTING.md#making-aspect-better).