@jdannberg 2024-06-04
I rebased to the current main, and I figured out how to properly use the tolerance of ARKode. I also added a test case that does not depend on resolution (the velocity is zero and it only runs in one cell) and has an analytical solution, to be better able to judge the accuracy. With these changes, the new version with SUNDIALs is more accurate, and it is also faster. The more the grain size changes within a time step, the more efficient SUNDIALs is compared to the old implementation. 

Here are some numbers from my laptop, run with the new test case I added, and changing the global refinements to 7 (just so that everything runs on lots of points and takes longer). I made 3 different test cases, one with the original reaction rate, one with 10x faster reactions, and one with 100x faster reaction (by changing the Grain growth rate constant). 
| | Total wallclock time elapsed since start | Assemble composition system |
|---|---|---|
|original growth rate |
|new | 6.07s | 1.6s |
|old | 8.01s | 2.48s |
|10x growth rate | | |
|new | 7.57s | 2.4s | 
|old | 13.7s | 5.24s | 
|100x growth rate | 
|new | 7.4s | 2.27s | 
|old | 17.5s | 7.11s |