@jdannberg 2024-06-05
I have addressed your comments and updated the test results. I went through all the tests one by one, and the new results all make sense, except for the grain_size_growth_log.prm test. This one clearly has wrong test results, but they were also wrong before, so it is not related to this change. This was an option we introduced before we could advect grain size on particles, so now it does not make sense to use it anymore. Therefore I would just completely remove that option in a follow-up PR.