@jdannberg 2024-06-06
@bangerth Can you have another look?

The last test that was not working was a cookbook, so it was an application model using grain size. The problem was that it was crashing because the initial step size passed into ARKode was too long. So I made this step size smaller, and I made sure that it does not affect the timing in the test case (ARKode increases the step size by a factor of 20 if it’s too small, so it would do at most 2 steps more if this initial step size was too small). However, how to pick this parameter likely depends on the model, so I’d like to make this an input parameter in a follow-up PR. 

The test output also changed, so I ran the whole cookbook again with both the old and new version, and with different time step sizes. The figures below shows that the version with ARKode is the more accurate one (it is closer to the version with the small time steps). So we know why the test changed and that the new version is right. 

![time_step_convergence](https://github.com/geodynamics/aspect/assets/7517347/42f1ebf3-c1b7-44f1-809b-4c5d21eb3c6e)

The new version also takes way longer for this application case (grain size is tracked on particles, and Particles: Update properties can take 10x as long with the new ARKode implementation). But maybe that's something we can try to fix as a follow-up as well.