@jdannberg 2024-06-06
Update: ARKode is slower in models with particles (as my application example) because they call the evaluate function once per particle, not per cell. In very simple problems, this can mean that the ARKode implementation is slower than the old one. However in my simple test cases the slowdown was never as dramatic (no matter what the grain growth rate was, it was at most 20%), so the dramatic slowdown in the application model must have been caused by something else.