@jdannberg 2024-06-06
I have squashed my commits and moved the `using VectorType = Vector<double>`. I have also added a changelog entry (since I had forgotten to do that).